Create base.ts and ensure index.ts is exporting modules correctly
